move demonstrator-group component into demonstrator/detail page
