+++ app/controllers/demonstrator.js
@@ -5,8 +5,8 @@ export default Ember.Controller.extend({
     this.set('selectedEvent', null);
     return this.set('currentView', null);
   },
-  subMenu: Ember.computed('model.eventTemplates', 'model.eventTemplates.[]', function() {
-    return this.get('model.eventTemplates').map(eventTemplate => ({
+  subMenu: Ember.computed('model', 'model.[]', function() {
+    return this.get('model').map(eventTemplate => ({
       key: eventTemplate,
       id: eventTemplate.get('id'),
       route: 'demonstrator.detail',
