+++ app/routes/index.js
@@ -1,6 +1,5 @@
 import Ember from 'ember';
 
-
 export default Ember.Route.extend({
   beforeModel: function() {
     return this.transitionTo('news');
