+++ app/routes/login-shibboleth.js
@@ -1,5 +1,4 @@
 import Ember from 'ember';
-import UnauthenticatedRouteMixin from 'ember-simple-auth/mixins/unauthenticated-route-mixin';
 import ErrorRouteMixin from '../mixins/error-route';
 
 export default Ember.Route.extend(ErrorRouteMixin,  {
