--- app/templates/components/demonstrator-group.hbs
@@ -1,9 +1,0 @@
-<div class="text label">
-  Feladatsor
-</div>
-<a class="button no-margin-top" {{action 'generateSheet'}}>generálás</a>
-<a class="button no-margin-top no-margin-left" {{action 'generateSheetSupplementary'}}>pót generálás</a>
-
-{{#each-in (await sortedEventsByCourseCode) as |courseCode events|}}
-  {{demonstrator-group-data currentEventTemplate=currentEventTemplate events=(await events) evaluateEvent=(action 'evaluateEvent')}}
-{{/each-in}}
