+++ app/templates/demonstrator/detail.hbs
@@ -1,1 +1,9 @@
-{{demonstrator-group currentEventTemplate=model evaluateEvent=(action 'evaluateEvent')}}
+<div class="text label">
+  Feladatsor
+</div>
+<a class="button no-margin-top" {{action 'generateSheet'}}>generálás</a>
+<a class="button no-margin-top no-margin-left" {{action 'generateSheetSupplementary'}}>pót generálás</a>
+
+{{#each-in (await sortedEventsByCourseCode) as |courseCode events|}}
+  {{demonstrator-group-data currentEventTemplate=model events=(await events) evaluateEvent=(action 'evaluateEvent')}}
+{{/each-in}}
